Added multifile reading
VM translator can now recieve and entire directory as an input. It will then read and translate all .vm files, starting with Sys.vm. The result is a .asm file in the selected directory which has the same name as the directory. The tkinter UI needs some work to improve usability.